move convert_alert_to_metric
